@@ -3,8 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "16.11.0",
+    "bootstrap": "^5.1.3",
+    "react": "^17.0.2",
+    "react-bootstrap": "^2.0.3",
     "react-dom": "16.11.0",
+    "react-router-dom": "^6.0.2",
     "react-scripts": "3.2.0",
     "web3": "1.2.2"
   },
